Provide dscl -url output support for OS X 10.4 clients using the directoryservice provider.
We refactored the directoryservice provider in puppet 0.24.7 to use the -plist option so that
output could be parsed more easily to resolve a bug with values with spaces in the name.
We missed that 10.4 does not support this flag, so this patch adds -url output support back in
for OS X 10.4 clients only as well as a new -url output parser that copes with spaces. 10.5
clients continue to use the -plist method.
Also includes some miscellaneous cleanup of methods and removal of extraneous comments.

Signed-off-by: Nigel Kersten <nigelk@google.com>

Don't use sw_vers to determine OS X versions, use the Facter value instead.

This relies upon another patch to ralsh to ensure that facts are loaded.